+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Improved variable names
 * Added SaveText() to Unload()
 * Added comments to important procedures
+* Added automatic reconnect
 * Removed Library dependencies (see 2.5.2)
 
 ## 2.5.6 - Improvements